// src/misc/enum.ts
// Export an enum called UserType with two values, Student and Mentor
export enum UserType {
  Student = "Student",
  Mentor = "Mentor",
}

// Export an enum called Department with ten values, representing different departments
export enum Department {
  IT = "IT",
  SoftwareEngineering = "Software Engineering",
  ComputerSciences = "Computer Sciences",
  ChemicalEngineering = "Chemical Engineering",
  AccountingFinance = "Accounting & Finance",
  Pharmacist = "Pharmacist",
  MechanicalEngineering = "Mechanical Engineering",
  ElectricalEngineering = "Electrical Engineering",
  CivilEngineering = "Civil Engineering",
  AerospaceEngineering = "Aerospace Engineering",
  BiomedicalEngineering = "Biomedical Engineering",
  EnvironmentalEngineering = "Environmental Engineering",
  Mathematics = "Mathematics",
  Physics = "Physics",
  Biology = "Biology",
  Chemistry = "Chemistry",
}

// Export an enum called ConnectionStatus with three values, Pending, Approved, and Rejected
export enum ConnectionStatus {
  Pending = "Pending",
  Approved = "Approved",
  Rejected = "Rejected",
}

// Export an enum called Gender with three values, Male, Female, and Other
export enum Gender {
  Male = "Male",
  Female = "Female",
  Other = "Other",
}